Completed the layout and form functions
Used CSS grid to layout the application.
Create a button that will open a dialog of
a form. In JS I was able to get the values of the input.
Next time I will need to work on a way to display
the new book that has been added in the library section.