Completed the addbook to library function
I was able to create a div with the books title,
author, pages and if it has been read.
It is organized by css grid. created a button for each
book. When the button is clicked it will remove the
div or the book but also remove it from the myLibrary array
based on its unique ID.
Next step is to add some CSS to make each div look like
a book.