Completed the library app
Created a function that will successfully
add the book to the library array. Also added a button
whichn will remove the book from the array and from
the screen. Also made the forms clear out once the user
submits  it. I did a bit of css for styling.
Next time around, I will try to refract the code
and get the css to be more appealing.